Repackaging the control frame
Including separate control frame into UID+RTG+low-level control, separate initialization of sensors serving for high-level controller and low-level controller
